# Swagger/swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters."
  version: "2.0.0"
  title: "Swagger Document"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "nhom1@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "petstore.swagger.io"
basePath: "/v2"
...
schemes:
  - "https"
  - "http"
paths:
  /Register:
    post:
      tags:
        - "Users"
      summary: "Add a new user"
      description: ""
      operationId: "addUser"
      consumes:
        - "application/json"
        - "application/xml"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Creat a new user"
          required: true
          schema:
            $ref: ""
      responses:
        "405":
          description: "Invalid input"
  /Login:
    post:
      tags:
        - "Users"
      summary: "Log in to the system"
      operationId: "Login User"
      consumes:
        - "application/json"
        - "application/xml"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "User object logged into the store"
          required: false
          schema:
            $ref: ""
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Product not found"
        "405":
          description: "Validation exception"
...